Recfactor class logic
