+++ qc-circuit/circuit.cpp
@@ -1,6 +1,7 @@
 #include "circuit.h"
 #include "operators.h"
 #include <cmath>
+#include <memory>
 #include <string>
 #include <vector>
 
@@ -23,47 +24,42 @@ Circuit::Circuit(int numero_qubits){
   this->numero_qubits = numero_qubits;
   for(long int i=0; i<std::pow(2,numero_qubits); i++){
     std::string str = number_to_bits(i, numero_qubits);
-    (this->estados).push_back(str);
     if(i==0){
-      (this->amplitudes).push_back(1.0);  
+      (this->base_computacional)[str] = 1.0;
     }else{
-      (this->amplitudes).push_back(0.0);
+      (this->base_computacional)[str] = 0.0;
     };
   };
 };
 
 
 void Circuit::set_estado_inicial(const std::vector< std::complex<double> > vector_estado){
-  int aux = 0;
-  for(auto string_estado: this->estados){
-    int index = std::stoi(string_estado, nullptr, 2);
-    (this->amplitudes)[aux] = vector_estado.at(index);
-    aux = aux+1;
+  for( auto it=(this->base_computacional).begin(); it!=(this->base_computacional).end(); it++ ){
+    int index = std::stoi( it->first, nullptr, 2 );
+    (this->base_computacional)[it->first] = vector_estado.at(index);
   };
 };
 
  
 std::vector< std::complex<double> > Circuit::obtener_estado(){
-  //Crear estado de 2^n
   std::vector< std::complex<double> > vector_estado(this->numero_qubits, 0.0);
-
-  //Iterar sobre los estados para obtener el indice y mover la probabilidad al estado
-  for( auto string_estado: this->estados ){
-    int index = std::stoi( string_estado, nullptr, 2 );
-    vector_estado[index] = (this->amplitudes).at(index);
+  
+  for( auto it=(this->base_computacional).begin(); it!=(this->base_computacional).end(); it++ ){
+    int index = std::stoi( it->first, nullptr, 2 );
+    vector_estado[index] = it->second;
   };
   return vector_estado;
 };
 
 
 //Toda la informacion de las wires y el operador esta el objeto
-void Circuit::aplicar_operador(const Operator operador){
-  (this->queue_operadores).push_back(operador);
+void Circuit::aplicar_operador( std::shared_ptr<Operator> op ){
+  (this->queue_operadores).push_back(op);
 };
 
 
 void Circuit::eliminar_operador(const int indice){
-  std::vector<Operator>::iterator it = (this->queue_operadores).begin() + indice;
+  auto it = (this->queue_operadores).begin() + indice;
   (this->queue_operadores).erase(it);
 };
 
@@ -72,5 +68,13 @@ void Circuit::mostrar_circuito(){
   std::cout << "wop" << std::endl;
 };
 
-void Circuit::ejecutar_circuit(){};
+void Circuit::ejecutar_circuito(){
+  //El operador se aplica a cada estado de la base computacional
+  for( auto op: (this->queue_operadores) ){
+     
+    for( auto it=(this->base_computacional).begin(); it!=(this->base_computacional).end(); it++ ){
+      auto resultado = op->aplicar_operador( it->first, it->second );
+    };
+  }
+};
